cloudsam/tests/unit/test_dynamo.py
import json
import pytest

import boto3

# Import the lambda_handler function from your code

from moto import mock_dynamodb2
from hello_world import app

# Define the DynamoDB table name
TABLE_NAME = 'cloudresume'

# Define the initial 'views' count for testing
#INITIAL_VIEWS_COUNT = 10

@mock_dynamodb2
def test_app():
    try:
    # Create a mock DynamoDB table
        dynamodb = boto3.resource('dynamodb', region_name='us-east-1')
        table = dynamodb.create_table(
            TableName=TABLE_NAME,
            KeySchema=[
                {
                'AttributeName': 'id',
                'KeyType': 'HASH'
                }
            ],
            AttributeDefinitions=[
                {
                    'AttributeName': 'id',
                    'AttributeType': 'S'
                }
            ],
            ProvisionedThroughput={
                'ReadCapacityUnits': 5,
                'WriteCapacityUnits': 5
            }
        )
        initial_views_count = 10
    # Populate the table with an item for testing
        table.put_item(
            Item={
               'id': '1',
                'views': initial_views_count
            }
        )
    # Invoke the lambda_handler function
        event = {}
        context = {}
        returned_views_count = app.get_item()
        print(returned_views_count)
# Check if the 'views' count is updated correctly
        expected_views_count = int(returned_views_count['views']) + 1
        updated_views_count = app.put_item(returned_views_count['views'])
        assert expected_views_count == updated_views_count, f"Expected views count: {expected_views_count}, Updated views count: {updated_views_count}"
        
        print("Test passed successfully: 200 OK")
    except AssertionError as e:
        print(f"Assertion Error: {e}")
        raise e
    except Exception as e:
        print(f"Test failed: {e}")
        raise e
    
    """  # Retrieve the item from the DynamoDB table to verify the update
    response = table.get_item(
        Key={
            'id': '1'
        }
    )
    updated_item = response['Item']
    assert updated_item['views'] == expected_views_count """
