cloudsam/tests/unit/requirements.txt
boto3==1.34.86
botocore==1.34.86
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
exceptiongroup==1.2.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
moto-ext==2.2.7
packaging==24.0
pluggy==1.4.0
pycparser==2.22
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
responses==0.25.0
s3transfer==0.10.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.18
Werkzeug==3.0.2
xmltodict==0.13.0
